for setting the permission part in side menu and displaying a message on deleting the grading period which is already assigned to some exam

+++ modules/exam/classes/controller/examgroup.php
@@ -7,6 +7,9 @@ class Controller_Examgroup extends Controller_Base {
     protected $_errors = array();
     
     public function action_index(){
+        
+        $msg = $this->request->param('msg');
+        
         if($this->request->param('sort')){
             $sort = $this->request->param('sort');
         } else {
@@ -85,6 +88,7 @@ class Controller_Examgroup extends Controller_Base {
                     ->bind('pagination', $pagination)
                     ->bind('filter_name', $filter_name)
                     ->bind('filter_url', $filter_url)
+                    ->bind('msg', $msg)
                     ;
         
         $this->content = $view; 
@@ -182,10 +186,18 @@ class Controller_Examgroup extends Controller_Base {
     public function action_delete(){
         if($this->request->method() === 'POST' && $this->request->post('selected')){
             foreach($this->request->post('selected') as $id){
-                ORM::factory('examgroup', $id)->delete();
+                $exam = ORM::factory('exam');
+                $exam->where('examgroup_id', '=', $id);
+                $count = $exam->count_all();
+                if($count > 0){
+                   $msg = 1;
+                } else {
+                    ORM::factory('examgroup', $id)->delete();
+                    $msg = 0;
+                }
             }
         }
-        Request::current()->redirect('examgroup');
+        Request::current()->redirect('examgroup/index/msg/'.$msg);
     }
 
     public function action_nil_exams() {
